clang-tidy: address google-runtime-int

+++ silkworm/infra/common/mem_usage.cpp
@@ -52,9 +52,9 @@ size_t get_mem_usage(bool resident) {
     size_t vm_size = 0, rm_size = 0;
     FILE* file = fopen("/proc/self/statm", "r");
     if (file) {
-        unsigned long vm = 0, rm = 0;
+        size_t vm = 0, rm = 0;
         // NOLINTNEXTLINE(cert-err34-c)
-        if (fscanf(file, "%lu %lu", &vm, &rm) == 2) {  // the first 2 num: vm size, resident set size
+        if (fscanf(file, "%zu %zu", &vm, &rm) == 2) {  // the first 2 num: vm size, resident set size
             vm_size = vm * static_cast<size_t>(getpagesize());
             rm_size = rm * static_cast<size_t>(getpagesize());
         }
